feat: local db 조회, 삭제 연동
